// frontend/src/api-generated/models/dangKyDonViRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DangKyDonViRequest {
  tenPhapLy: string;
  tenMienHeThong: string;
  maSoThue?: string;
  tenNguoiDaiDien: string;
  tenDangNhapAdmin: string;
  matKhauAdmin: string;
  tenAdmin: string;
  emailAdmin: string;
}